federated-secure-fl/server/requirements.txt
flwr==1.8.0
fastapi
uvicorn
python-jose[cryptography]
httpx
sqlalchemy
python-dotenv
numpy
